Implement role and permission management features, including role deletion checks, permission assignment, and a dedicated permissions view

+++ app/Http/Controllers/RoleController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Http\Requests\RoleRequest;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Validator;
+use Spatie\Permission\Models\Permission;
 use Spatie\Permission\Models\Role;
 
 class RoleController extends Controller
@@ -13,7 +14,7 @@ class RoleController extends Controller
     {
         return view('admin.roles.index', [
             'title' => 'Roles',
-            'items' => Role::all(),
+            'items' => Role::where('name', '!=', 'SUPER ADMIN')->get(),
         ]);
     }
 
@@ -55,10 +56,49 @@ class RoleController extends Controller
     public function destroy(Role $role)
     {
         try {
+            $role = Role::WithCount('users')->findOrFail($role->id);
+            if ($role->users_count > 0) {
+                return to_route('roles.index')->with('error', 'No se puede eliminar el rol porque está asignado a uno o más usuarios.');
+            }
             $role->delete();
             return to_route('roles.index')->with('success', 'Rol eliminado exitosamente.');
         } catch (\Exception $e) {
             return to_route('roles.index')->with('error', 'Error al eliminar el rol: ' . $e->getMessage());
+        }
+    }
+
+    public function showPermissions(Role $role)
+    {
+        $permissions = Permission::all()->groupBy(function($permission){
+            if(stripos($permission->name, 'setting') !== false){return 'Configuraciones';}
+            if(stripos($permission->name, 'role') !== false){return 'Roles';}
+            if(stripos($permission->name, 'user') !== false){return 'Usuarios';}
+            if(stripos($permission->name, 'space') !== false){return 'Espacios de Estacionamiento';}
+            if(stripos($permission->name, 'rate') !== false){return 'Tarifas';}
+            if(stripos($permission->name, 'customer') !== false){return 'Clientes';}
+            if(stripos($permission->name, 'vehicle') !== false){return 'Vehículos';}
+            if(stripos($permission->name, 'ticket') !== false){return 'Tickets';}
+            if(stripos($permission->name, 'invoice') !== false){return 'Facturaciones';}
+            if(stripos($permission->name, 'analytic') !== false){return 'Análisis y Gráficos';}
+            if(stripos($permission->name, 'report') !== false){return 'Reportes';}
+        });
+
+        return view('admin.roles.show_permissions', [
+            'title' => 'Asignar Permisos al Rol: ' . $role->name,
+            'role' => $role,
+            'permissions' => $permissions,
+        ]);
+    }
+
+    public function assignPermissions(Request $request, Role $role)
+    {
+        try {
+            $role = Role::findOrFail($role->id);
+            $role->permissions()->sync($request->permissions);
+
+            return to_route('roles.index')->with('success', 'Permisos asignados exitosamente.');
+        } catch (\Exception $e) {
+            return to_route('roles.index')->with('error', 'Error al asignar permisos: ' . $e->getMessage());
         }
     }
 }
